+++ packages/firebase_auth/ios/Classes/FirebaseAuthPlugin.m
@@ -176,6 +176,28 @@ int nextHandle = 0;
                                         completion:^(FIRUser *user, NSError *error) {
                                           [self sendResult:result forUser:user error:error];
                                         }];
+  } else if ([@"linkWithTwitterCredential" isEqualToString:call.method]) {
+    NSString *authToken = call.arguments[@"authToken"];
+    NSString *authTokenSecret = call.arguments[@"authTokenSecret"];
+    FIRAuthCredential *credential =
+        [FIRTwitterAuthProvider credentialWithToken:authToken secret:authTokenSecret];
+    [[FIRAuth auth].currentUser linkWithCredential:credential
+                                        completion:^(FIRUser *user, NSError *error) {
+                                          [self sendResult:result forUser:user error:error];
+                                        }];
+  } else if ([@"updateEmail" isEqualToString:call.method]) {
+    NSString *email = call.arguments[@"email"];
+    [[FIRAuth auth].currentUser updateEmail:email
+                                 completion:^(NSError *error) {
+                                   [self sendResult:result forUser:nil error:error];
+                                 }];
+  } else if ([@"updatePassword" isEqualToString:call.method]) {
+    NSString *password = call.arguments[@"password"];
+    [[FIRAuth auth].currentUser updatePassword:password
+                                    completion:^(NSError *error) {
+                                      [self sendResult:result forUser:nil error:error];
+                                    }];
+
   } else if ([@"updateProfile" isEqualToString:call.method]) {
     FIRUserProfileChangeRequest *changeRequest = [[FIRAuth auth].currentUser profileChangeRequest];
     if (call.arguments[@"displayName"]) {
@@ -187,12 +209,6 @@ int nextHandle = 0;
     [changeRequest commitChangesWithCompletion:^(NSError *error) {
       [self sendResult:result forUser:nil error:error];
     }];
-  } else if ([@"updateEmail" isEqualToString:call.method]) {
-    NSString *toEmail = call.arguments[@"email"];
-    [[FIRAuth auth].currentUser updateEmail:toEmail
-                                 completion:^(NSError *_Nullable error) {
-                                   [self sendResult:result forUser:nil error:error];
-                                 }];
   } else if ([@"signInWithCustomToken" isEqualToString:call.method]) {
     NSString *token = call.arguments[@"token"];
     [[FIRAuth auth] signInWithCustomToken:token
@@ -275,7 +291,13 @@ int nextHandle = 0;
   for (id<FIRUserInfo> userInfo in user.providerData) {
     [providerData addObject:toDictionary(userInfo)];
   }
+
+  long creationDate = [user.metadata.creationDate timeIntervalSince1970];
+  long lastSignInDate = [user.metadata.lastSignInDate timeIntervalSince1970];
+
   NSMutableDictionary *userData = [toDictionary(user) mutableCopy];
+  userData[@"creationTimestamp"] = [NSNumber numberWithLong:creationDate];
+  userData[@"lastSignInTimestamp"] = [NSNumber numberWithInt:lastSignInDate];
   userData[@"isAnonymous"] = [NSNumber numberWithBool:user.isAnonymous];
   userData[@"isEmailVerified"] = [NSNumber numberWithBool:user.isEmailVerified];
   userData[@"providerData"] = providerData;