+++ packages/firebase_messaging/lib/firebase_messaging.dart
@@ -22,7 +22,7 @@ class FirebaseMessaging {
       : _channel = channel,
         _platform = platform;
 
-  static final FirebaseMessaging _instance = new FirebaseMessaging.private(
+  static final FirebaseMessaging _instance = FirebaseMessaging.private(
       const MethodChannel('plugins.flutter.io/firebase_messaging'),
       const LocalPlatform());
 
@@ -48,7 +48,7 @@ class FirebaseMessaging {
   }
 
   final StreamController<IosNotificationSettings> _iosSettingsStreamController =
-      new StreamController<IosNotificationSettings>.broadcast();
+      StreamController<IosNotificationSettings>.broadcast();
 
   /// Stream that fires when the user changes their notification settings.
   ///
@@ -71,7 +71,7 @@ class FirebaseMessaging {
   }
 
   final StreamController<String> _tokenStreamController =
-      new StreamController<String>.broadcast();
+      StreamController<String>.broadcast();
 
   /// Fires when a new FCM token is generated.
   Stream<String> get onTokenRefresh {
@@ -80,9 +80,7 @@ class FirebaseMessaging {
 
   /// Returns the FCM token.
   Future<String> getToken() {
-    return _token != null
-        ? new Future<String>.value(_token)
-        : onTokenRefresh.first;
+    return _token != null ? Future<String>.value(_token) : onTokenRefresh.first;
   }
 
   /// Subscribe to topic in background.
@@ -108,7 +106,7 @@ class FirebaseMessaging {
         }
         return null;
       case "onIosSettingsRegistered":
-        _iosSettingsStreamController.add(new IosNotificationSettings._fromMap(
+        _iosSettingsStreamController.add(IosNotificationSettings._fromMap(
             call.arguments.cast<String, bool>()));
         return null;
       case "onMessage":
@@ -118,16 +116,12 @@ class FirebaseMessaging {
       case "onResume":
         return _onResume(call.arguments.cast<String, dynamic>());
       default:
-        throw new UnsupportedError("Unrecognized JSON message");
+        throw UnsupportedError("Unrecognized JSON message");
     }
   }
 }
 
 class IosNotificationSettings {
-  final bool sound;
-  final bool alert;
-  final bool badge;
-
   const IosNotificationSettings({
     this.sound = true,
     this.alert = true,
@@ -138,6 +132,10 @@ class IosNotificationSettings {
       : sound = settings['sound'],
         alert = settings['alert'],
         badge = settings['badge'];
+
+  final bool sound;
+  final bool alert;
+  final bool badge;
 
   @visibleForTesting
   Map<String, dynamic> toMap() {