+++ packages/google_maps_flutter/example/lib/main.dart
@@ -8,12 +8,14 @@ import 'map_ui.dart';
 import 'move_camera.dart';
 import 'page.dart';
 import 'place_marker.dart';
+import 'scrolling_map.dart';
 
 final List<Page> _allPages = <Page>[
   MapUiPage(),
   AnimateCameraPage(),
   MoveCameraPage(),
   PlaceMarkerPage(),
+  ScrollingMapPage(),
 ];
 
 class MapsDemo extends StatelessWidget {
@@ -42,6 +44,5 @@ class MapsDemo extends StatelessWidget {
 }
 
 void main() {
-  final List<NavigatorObserver> observers = <NavigatorObserver>[];
-  runApp(MaterialApp(home: MapsDemo(), navigatorObservers: observers));
+  runApp(MaterialApp(home: MapsDemo()));
 }