+++ packages/google_maps_flutter/lib/src/ui.dart
@@ -71,6 +71,23 @@ class MinMaxZoomPreference {
 /// When used to change configuration, null values will be interpreted as
 /// "do not change this configuration option".
 class GoogleMapOptions {
+  /// Creates a set of map user interface configuration options.
+  ///
+  /// By default, every non-specified field is null, meaning no desire to change
+  /// user interface defaults or current configuration.
+  GoogleMapOptions({
+    this.cameraPosition,
+    this.compassEnabled,
+    this.cameraTargetBounds,
+    this.mapType,
+    this.minMaxZoomPreference,
+    this.rotateGesturesEnabled,
+    this.scrollGesturesEnabled,
+    this.tiltGesturesEnabled,
+    this.trackCameraPosition,
+    this.zoomGesturesEnabled,
+  });
+
   /// The desired position of the map camera.
   ///
   /// This field is used to indicate initial camera position and to update that
@@ -108,23 +125,6 @@ class GoogleMapOptions {
   /// True if the map view should respond to zoom gestures.
   final bool zoomGesturesEnabled;
 
-  /// Creates a set of map user interface configuration options.
-  ///
-  /// By default, every non-specified field is null, meaning no desire to change
-  /// user interface defaults or current configuration.
-  GoogleMapOptions({
-    this.cameraPosition,
-    this.compassEnabled,
-    this.cameraTargetBounds,
-    this.mapType,
-    this.minMaxZoomPreference,
-    this.rotateGesturesEnabled,
-    this.scrollGesturesEnabled,
-    this.tiltGesturesEnabled,
-    this.trackCameraPosition,
-    this.zoomGesturesEnabled,
-  });
-
   /// Default user interface options.
   ///
   /// Specifies a map view that
@@ -159,7 +159,7 @@ class GoogleMapOptions {
     if (change == null) {
       return this;
     }
-    return new GoogleMapOptions(
+    return GoogleMapOptions(
       cameraPosition: change.cameraPosition ?? cameraPosition,
       compassEnabled: change.compassEnabled ?? compassEnabled,
       cameraTargetBounds: change.cameraTargetBounds ?? cameraTargetBounds,