+++ packages/firebase_analytics/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.4
+
+* Bump Android dependencies to latest.
+
 ## 1.0.3
 
 * Updated test and mockito dependencies to pick up Dart 2 support